VerifiedPrivileges. privileged (dipendenti)

+++ app/Http/Controllers/general_edit_controller.php
@@ -10,19 +10,18 @@ use DB;
 class general_edit_controller extends Controller
 {
     public static function address_star($id){
-        $uid=Auth::user()->id;
+        if(Auth::check()) {
+            $uid=Auth::user()->id;
         DB::table('users')->where('id', $uid)->update(['address_id' => $id]);
 
         return redirect('/profile#spedizione');
-
-
+        } else redirect('login');
     }
 
     public function delete_user_address($id){
-        Address::destroy($id);
-
-        return redirect('/profile#spedizione');
-
-
+        if(Auth::check()) {
+            Address::destroy($id);
+            return redirect('/profile#spedizione');
+        } else redirect('login');
     }
 }
